pyproject.toml
[project]
name = "spectre-rewrite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "emoji>=2.14.1",
    "ollama>=0.5.1",
    "sqlalchemy>=2.0.41",
]
